+++ src/constants/code-editor.ts
@@ -1,5 +1,5 @@
 import { SelectableValue } from '@grafana/data';
-import { CodeEditorSuggestionItem, CodeEditorSuggestionItemKind } from '@grafana/ui';
+import { CodeEditorSuggestionItem } from '@grafana/ui';
 
 import { CodeEditorType, CodeLanguage } from '../types';
 import {
@@ -85,116 +85,18 @@ export const CODE_LANGUAGE_OPTIONS: SelectableValue[] = [
 ];
 
 /**
- * Request Code Suggestions
- */
-const REQUEST_CODE_SUGGESTIONS: CodeEditorSuggestionItem[] = [
-  {
-    label: 'options',
-    kind: CodeEditorSuggestionItemKind.Property,
-    detail: "Panels' options.",
-  },
-  {
-    label: 'data',
-    kind: CodeEditorSuggestionItemKind.Property,
-    detail: 'Result set of panel queries.',
-  },
-  {
-    label: 'response',
-    kind: CodeEditorSuggestionItemKind.Property,
-    detail: "Request's response",
-  },
-  {
-    label: 'initial',
-    kind: CodeEditorSuggestionItemKind.Property,
-    detail: 'Parsed values from the Initial Request.',
-  },
-  {
-    label: 'elements',
-    kind: CodeEditorSuggestionItemKind.Property,
-    detail: 'Form Elements.',
-  },
-  {
-    label: 'locationService',
-    kind: CodeEditorSuggestionItemKind.Property,
-    detail: 'Allows to work with browser location and history.',
-  },
-  {
-    label: 'templateService',
-    kind: CodeEditorSuggestionItemKind.Property,
-    detail: 'Provides access to variables and allows to update Time Range.',
-  },
-  {
-    label: 'onOptionsChange',
-    kind: CodeEditorSuggestionItemKind.Method,
-    detail: 'Change handler to refresh panel.',
-  },
-  {
-    label: 'initialRequest',
-    kind: CodeEditorSuggestionItemKind.Method,
-    detail: 'Perform the Initial Request to reload panel.',
-  },
-  {
-    label: 'setInitial',
-    kind: CodeEditorSuggestionItemKind.Method,
-    detail: 'Allows to specify the initial values.',
-  },
-  {
-    label: 'notifySuccess',
-    kind: CodeEditorSuggestionItemKind.Method,
-    detail: 'Display successful notification.',
-  },
-  {
-    label: 'notifyError',
-    kind: CodeEditorSuggestionItemKind.Method,
-    detail: 'Display error notification.',
-  },
-  {
-    label: 'notifyWarning',
-    kind: CodeEditorSuggestionItemKind.Method,
-    detail: 'Display warning notification.',
-  },
-  {
-    label: 'toDataQueryResponse',
-    kind: CodeEditorSuggestionItemKind.Method,
-    detail: 'Parse the results from /api/ds/query.',
-  },
-];
-
-/**
  * Suggestions
  */
 export const CODE_EDITOR_SUGGESTIONS: Record<CodeEditorType, CodeEditorSuggestionItem[]> = {
   /**
    * Initial, update or reset request
    */
-  [CodeEditorType.REQUEST]: [
-    ...REQUEST_CODE_SUGGESTIONS,
-
-    /**
-     * Context
-     */
-    ...requestCodeParameters.suggestions,
-  ],
+  [CodeEditorType.REQUEST]: requestCodeParameters.suggestions,
 
   /**
    * Get data source payload
    */
-  [CodeEditorType.GET_PAYLOAD]: [
-    {
-      label: 'elements',
-      kind: CodeEditorSuggestionItemKind.Property,
-      detail: 'Form Elements.',
-    },
-    {
-      label: 'initial',
-      kind: CodeEditorSuggestionItemKind.Property,
-      detail: 'Parsed values from the Initial Request.',
-    },
-    /**
-     * Context
-     */
-    ...getPayloadCodeParameters.suggestions,
-  ],
+  [CodeEditorType.GET_PAYLOAD]: getPayloadCodeParameters.suggestions,
 
   /**
    * Element value changed
@@ -204,68 +106,15 @@ export const CODE_EDITOR_SUGGESTIONS: Record<CodeEditorType, CodeEditorSuggestio
   /**
    * Element disable if
    */
-  [CodeEditorType.ELEMENT_DISABLE_IF]: [
-    {
-      label: 'elements',
-      kind: CodeEditorSuggestionItemKind.Property,
-      detail: 'Form Elements.',
-    },
-    {
-      label: 'replaceVariables',
-      kind: CodeEditorSuggestionItemKind.Method,
-      detail: 'Interpolate variables.',
-    },
-
-    /**
-     * Context
-     */
-    ...disableIfCodeParameters.suggestions,
-  ],
+  [CodeEditorType.ELEMENT_DISABLE_IF]: disableIfCodeParameters.suggestions,
 
   /**
    * Element show if
    */
-  [CodeEditorType.ELEMENT_SHOW_IF]: [
-    {
-      label: 'elements',
-      kind: CodeEditorSuggestionItemKind.Property,
-      detail: 'Form Elements.',
-    },
-    {
-      label: 'replaceVariables',
-      kind: CodeEditorSuggestionItemKind.Method,
-      detail: 'Interpolate variables.',
-    },
-
-    /**
-     * Context
-     */
-    ...showIfCodeParameters.suggestions,
-  ],
+  [CodeEditorType.ELEMENT_SHOW_IF]: showIfCodeParameters.suggestions,
 
   /**
    * Element get options
    */
-  [CodeEditorType.ELEMENT_GET_OPTIONS]: [
-    {
-      label: 'data',
-      kind: CodeEditorSuggestionItemKind.Property,
-      detail: 'Result set of panel queries.',
-    },
-    {
-      label: 'elements',
-      kind: CodeEditorSuggestionItemKind.Property,
-      detail: 'Form Elements.',
-    },
-    {
-      label: 'replaceVariables',
-      kind: CodeEditorSuggestionItemKind.Method,
-      detail: 'Interpolate variables.',
-    },
-
-    /**
-     * Context
-     */
-    ...getOptionsCodeParameters.suggestions,
-  ],
+  [CodeEditorType.ELEMENT_GET_OPTIONS]: getOptionsCodeParameters.suggestions,
 };