+++ src/utils/form-element.ts
@@ -242,11 +242,9 @@ export const toLocalFormElement = (element: FormElement): LocalFormElement => {
 
   let showIfFn: ShowIfHelper = () => true;
   if (showIf || showIf?.trim()) {
-    const fn = new Function('elements', 'replaceVariables', 'context', showIf);
+    const fn = new Function('context', showIf);
     showIfFn = ({ elements, replaceVariables }: { elements: FormElement[]; replaceVariables: InterpolateFunction }) =>
       fn(
-        elements,
-        replaceVariables,
         showIfCodeParameters.create({
           panel: {
             elements,
@@ -262,7 +260,7 @@ export const toLocalFormElement = (element: FormElement): LocalFormElement => {
 
   let disableIfFn: DisableIfHelper = () => false;
   if (disableIf || disableIf?.trim()) {
-    const fn = new Function('elements', 'replaceVariables', 'context', disableIf);
+    const fn = new Function('context', disableIf);
     disableIfFn = ({
       elements,
       replaceVariables,
@@ -271,8 +269,6 @@ export const toLocalFormElement = (element: FormElement): LocalFormElement => {
       replaceVariables: InterpolateFunction;
     }) =>
       fn(
-        elements,
-        replaceVariables,
         disableIfCodeParameters.create({
           panel: {
             elements,
@@ -317,7 +313,7 @@ export const toLocalFormElement = (element: FormElement): LocalFormElement => {
         }));
       };
     } else if (element.optionsSource === OptionsSource.CODE) {
-      const fn = new Function('data', 'elements', 'replaceVariables', 'context', element.getOptions || 'return []');
+      const fn = new Function('context', element.getOptions || 'return []');
       getOptions = ({
         data,
         elements,
@@ -328,9 +324,6 @@ export const toLocalFormElement = (element: FormElement): LocalFormElement => {
         replaceVariables: InterpolateFunction;
       }) =>
         fn(
-          data,
-          elements,
-          replaceVariables,
           getOptionsCodeParameters.create({
             panel: {
               data,