+++ CHANGELOG.md
@@ -1,6 +1,39 @@
 # Change Log
 
-## 3.8.0 (IN PROGRESS)
+## 4.0.0 (IN PROGRESS)
+
+### Breaking changes
+
+- Requires Grafana 10 and Grafana 11
+- Data Source requests were updated to use Query Editor.
+- Removed non-context code parameters. Please update parameters to use `context`.
+
+### Code parameters migration guide
+
+- data -> context.panel.data
+- elements -> context.panel.elements
+- initial -> context.panel.initial
+- initialRequest -> context.panel.initialRequest
+- locationService -> context.grafana.locationService
+- notifyError -> context.grafana.notifyError
+- notifySuccess -> context.grafana.notifySuccess
+- notifyWarning -> context.grafana.notifyWarning
+- onChange -> context.panel.onChange
+- onOptionsChange -> context.panel.onOptionsChange
+- options -> context.panel.options
+- replaceVariables -> context.grafana.replaceVariables
+- response -> context.panel.response
+- setInitial -> context.panel.setInitial
+- templateService -> context.grafana.templateService
+- toDataQueryResponse -> context.utils.toDataQueryResponse
+
+### Features / Enhancements
+
+- Updated name to Business Forms Panel (#361)
+- Add support frontend data sources (#361)
+- Prepare for Grafana 11 (#399)
+
+## 3.8.0 (2024-05-30)
 
 ### Features / Enhancements
 
@@ -11,6 +44,7 @@
 - Added input type time (#385)
 - Updated hiding/show multi-select element (#389)
 - Updated allowed files for file type (#388)
+- Added Form validation provisioning dashboard (#411)
 
 ## 3.7.0 (2024-03-10)